+++ test/jest/__mock__/stripesCore.mock.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+jest.mock('@folio/stripes/core', () => ({
+  stripesConnect: jest.fn((component) => component),
+}));